+++ src/Features/WagtailApi/index.ts
@@ -1,1 +1,2 @@
 export { WagtailBlock } from "./WagtailBlock"
+export { getWagtailPageById, wagtailQueryKey, usePage } from "./hooks"