+++ src/Pages/About/index.spec.tsx
@@ -1,8 +1,0 @@
-import { shallow } from 'enzyme'
-import * as React from 'react'
-
-import AboutPage from '.'
-
-it('renders without crashing', () => {
-    shallow(<AboutPage />)
-})