+++ src/reducer.ts
@@ -1,9 +1,6 @@
 /**
  * Combined redux reducer for all of the features.
  */
-
-import { connectRouter, LocationChangeAction } from "connected-react-router"
-import { History } from "history"
 import { combineReducers } from "redux"
 
 import { Action } from "./actions"
@@ -14,8 +11,6 @@ import { unitReducer as unit } from "Features/Units"
 /**
  * Redux reducer
  */
-export default (history: History) =>
-  combineReducers<StoreState, Action & LocationChangeAction>({
-    router: connectRouter(history),
-    unit,
-  })
+export default combineReducers<StoreState, Action>({
+  unit,
+})