+++ src/store.ts
@@ -1,7 +1,5 @@
 import * as Sentry from "@sentry/react"
-import { routerMiddleware } from "connected-react-router"
-import { createBrowserHistory } from "history"
-import { applyMiddleware, createStore, Store } from "redux"
+import { createStore, Store } from "redux"
 import { composeWithDevTools } from "redux-devtools-extension"
 
 import rootReducer from "./reducer"
@@ -11,16 +9,13 @@ import { StoreState } from "./Shared/constants/store"
 /**
  * Browser history
  */
-export const history = createBrowserHistory()
+// export const history = createBrowserHistory()
 
 const sentryReduxEnhancer = Sentry.createReduxEnhancer({})
 
 /**
  * Global Redux store
  */
-export const store: Store<StoreState> = createStore(
-  rootReducer(history),
-  composeWithDevTools(applyMiddleware(routerMiddleware(history)), sentryReduxEnhancer)
-)
+export const store: Store<StoreState> = createStore(rootReducer, composeWithDevTools(sentryReduxEnhancer))
 
 export default store